Fix partition tests in storage API tests
- Partitions tests moved to different directory
- Added support of normal disk (not partition) as TEST_STORAGE_PATH
