+++ openvair/modules/storage/tests/api/localfs/test_create_localfs.py
@@ -1,7 +1,7 @@
 """Integration tests for storage creation endpoints (localfs).
 
 Covers:
-- Successful storage (and local disk partition) creation.
+- Successful storage creation.
 - Validation errors (missing fields, invalid name, specs, etc.).
 - Logical errors (nonexistent path).
 - Unauthorized access.
@@ -11,7 +11,6 @@ from fastapi import status
 from fastapi.testclient import TestClient
 
 from openvair.libs.testing.utils import (
-    get_disk_partitions,
     wait_for_field_value,
     generate_test_entity_name,
 )
@@ -223,82 +222,4 @@ def test_create_storage_unauthorized(
     }
 
     response = unauthorized_client.post('/storages/create/', json=storage_data)
-    assert response.status_code == status.HTTP_401_UNAUTHORIZED
-
-
-def test_create_local_partition_success(
-    client: TestClient, target_disk_path: str
-) -> None:
-    """Test successful creation of local disk partition."""
-    partition_data = {
-        'local_disk_path': target_disk_path,
-        'storage_type': 'local_partition',
-        'size_value': 10 * 1000 * 1000,  # 10 MB
-        'size_unit': 'B',
-    }
-
-    response = client.post(
-        '/storages/local-disks/create_partition/', json=partition_data
-    )
-    assert response.status_code == status.HTTP_200_OK
-    result = response.json()
-    assert 'path' in result
-    assert 'size' in result
-    assert result['parent'] == target_disk_path
-    assert result['size'] > 0
-    assert result['path'].startswith(target_disk_path)
-    partition_number = result['path'].replace(target_disk_path, '')
-    assert partition_number in get_disk_partitions(target_disk_path)
-
-
-def test_create_local_partition_nonexistent_disk(client: TestClient) -> None:
-    """Test partition creation on nonexistent disk."""
-    partition_data = {
-        'local_disk_path': '/nonexistent/disk',
-        'storage_type': 'local_partition',
-        'size_value': 10 * 1000 * 1000,  # 10 MB
-        'size_unit': 'B',
-    }
-
-    response = client.post(
-        '/storages/local-disks/create_partition/', json=partition_data
-    )
-    assert response.status_code == status.HTTP_500_INTERNAL_SERVER_ERROR
-    assert 'does not exist' in response.text.lower()
-
-
-def test_create_local_partition_insufficient_space(
-    client: TestClient, target_disk_path: str
-) -> None:
-    """Test partition creation with insufficient disk space."""
-    partition_data = {
-        'local_disk_path': target_disk_path,
-        'storage_type': 'local_partition',
-        'size_value': 10 * 1000 * 1000 * 1000 * 1000,  # 10 TB
-        'size_unit': 'B',
-    }
-
-    response = client.post(
-        '/storages/local-disks/create_partition/', json=partition_data
-    )
-    assert response.status_code in [
-        status.HTTP_422_UNPROCESSABLE_ENTITY,
-        status.HTTP_500_INTERNAL_SERVER_ERROR,
-    ]
-
-
-def test_create_local_partition_unauthorized(
-    unauthorized_client: TestClient,
-) -> None:
-    """Test unauthorized partition creation."""
-    partition_data = {
-        'local_disk_path': '/some/path',
-        'storage_type': 'local_partition',
-        'size_value': 10 * 1000 * 1000,  # 10 MB
-        'size_unit': 'B',
-    }
-
-    response = unauthorized_client.post(
-        '/storages/local-disks/create_partition/', json=partition_data
-    )
     assert response.status_code == status.HTTP_401_UNAUTHORIZED
